Scheduling and tidy up
Implemented scheduling, including adding and editing schedules and binding them to a device.
Fixed creation of config file so that a blank template is loaded if one does not exist.
Moved Add[Device/Schedule] and Edit[Device/Schedule] into [Device/Schedule]Manager. The states can be initiated with "edit" in their name for edit, or not for add (default).